тест для списка жанров; обновила тесты для app и main
