+++ src/components/genres-list/genres-list.test.js
@@ -1,11 +1,7 @@
 import React from "react";
-import {Provider} from "react-redux";
-import configureStore from "redux-mock-store";
 import renderer from "react-test-renderer";
-import App from "./app.jsx";
-import {Genres} from "../../consts";
-
-const mockStore = configureStore([]);
+import {GenresList} from "./genres-list.jsx";
+import {Genres} from "../../consts.js";
 
 const films = [
   {
@@ -26,7 +22,7 @@ const films = [
     id: 3,
     name: `The Autopsy of Jane Doe`,
     genre: `thriller`,
-    poster: ``,
+    poster: `https://unsplash.it/280/175/`,
     preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
   },
   {
@@ -45,24 +41,18 @@ const films = [
   },
 ];
 
-it(`Render App`, () => {
-  const store = mockStore({
-    genre: Genres.ALL,
-    films,
-  });
-
-  const tree = renderer
-    .create(
-        <Provider store={store}>
-          <App
-            films={films}
-          />
-        </Provider>, {
-          createNodeMock: () => {
-            return {};
-          }
+it(`Should GenresList render correctly`, () => {
+  const tree = renderer.create(
+      <GenresList
+        films={films}
+        genre={Genres.ALL}
+        changeGenre={() => {}}
+      />, {
+        createNodeMock: () => {
+          return {};
         }
-    ).toJSON();
+      }
+  ).toJSON();
 
   expect(tree).toMatchSnapshot();
 });
